This is the first commit.
This is a simpe game in its budding stage.

First the user fills the 10 X 10 field with the ships he is given while entering their oordinates.Then, the player plays the battleship game on his board (this game is in its budding stage. This is it for now) by attacking on a specific cell by entering its coordianes.